+++ webpack.config.js
@@ -1,0 +1,55 @@
+const path = require('path');
+const webpack = require('webpack');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const WebpackMd5Hash = require('webpack-md5-hash');
+module.exports = {
+    entry: { 
+        main: './src/index.js' 
+    },
+    output: {
+        filename: '[name].[chunkhash].js',
+        path: path.resolve(__dirname, './dist')
+    },
+    module: {
+        rules: [{
+                test: /\.js$/,
+                use: {loader: "babel-loader"},
+                exclude: /node_modules/
+            },
+            {
+                test: /\.css$/,
+                use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader']
+            },
+            {
+                test: /\.(png|jpg|gif|ico|svg)$/,
+                use: [
+                     'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+                     {
+                         loader: 'image-webpack-loader'
+                         
+                     },
+                    ],
+            },
+            {
+                test: /\.(eot|ttf|woff|woff2)$/,
+                loader: 'file-loader?name=./vendor/[name].[ext]'
+            }
+        ]
+    },
+    plugins: [ 
+        new MiniCssExtractPlugin({
+            filename: 'style.[contenthash].css'
+        }),
+        new HtmlWebpackPlugin({
+            inject: false,
+            hash: true,
+            template: './src/index.html',
+            filename: 'index.html'
+          }),
+        new WebpackMd5Hash(),
+        new webpack.DefinePlugin({
+            'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
+       })
+    ]
+}
